create music service: extract music logic into own class
